--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,13 @@
 module github.com/cespare/mph
 
-go 1.17
+go 1.18
+
+require (
+	github.com/zeebo/xxh3 v1.0.2
+	golang.org/x/exp v0.0.0-20230118134722-a68e582fa157
+)
+
+require (
+	github.com/klauspost/cpuid/v2 v2.2.3 // indirect
+	golang.org/x/sys v0.4.0 // indirect
+)