--- a/mph.go
+++ b/mph.go
@@ -1,7 +1,10 @@
 // Package mph implements a minimal perfect hash table over strings.
 package mph
 
-import "sort"
+import (
+	"github.com/zeebo/xxh3"
+	"golang.org/x/exp/slices"
+)
 
 // A Table is an immutable hash table that provides constant-time lookups of key
 // indices using a minimal perfect hash.
@@ -22,10 +25,9 @@ func Build(keys []string) *Table {
 		level1        = make([]uint32, nextPow2(len(keys)))
 		level1Mask    = len(level1) - 1
 		sparseBuckets = make([][]int, len(level0))
-		zeroSeed      = murmurSeed(0)
 	)
 	for i, s := range keys {
-		n := int(zeroSeed.hash(s)) & level0Mask
+		n := int(xxh3.HashStringSeed(s, 0)) & level0Mask
 		sparseBuckets[n] = append(sparseBuckets[n], i)
 	}
 	var buckets []indexBucket
@@ -34,16 +36,18 @@ func Build(keys []string) *Table {
 			buckets = append(buckets, indexBucket{n, vals})
 		}
 	}
-	sort.Sort(bySize(buckets))
+	slices.SortFunc(buckets, func(a, b indexBucket) bool {
+		return len(a.vals) > len(b.vals)
+	})
 
 	occ := make([]bool, len(level1))
 	var tmpOcc []int
 	for _, bucket := range buckets {
-		var seed murmurSeed
+		var seed uint32
 	trySeed:
 		tmpOcc = tmpOcc[:0]
 		for _, i := range bucket.vals {
-			n := int(seed.hash(keys[i])) & level1Mask
+			n := int(xxh3.HashStringSeed(keys[i], uint64(seed))) & level1Mask
 			if occ[n] {
 				for _, n := range tmpOcc {
 					occ[n] = false
@@ -77,9 +81,9 @@ func nextPow2(n int) int {
 
 // Lookup searches for s in t and returns its index and whether it was found.
 func (t *Table) Lookup(s string) (n uint32, ok bool) {
-	i0 := int(murmurSeed(0).hash(s)) & t.level0Mask
+	i0 := int(xxh3.HashStringSeed(s, 0)) & t.level0Mask
 	seed := t.level0[i0]
-	i1 := int(murmurSeed(seed).hash(s)) & t.level1Mask
+	i1 := int(xxh3.HashStringSeed(s, uint64(seed))) & t.level1Mask
 	n = t.level1[i1]
 	return n, s == t.keys[int(n)]
 }
@@ -88,9 +92,3 @@ type indexBucket struct {
 	n    int
 	vals []int
 }
-
-type bySize []indexBucket
-
-func (s bySize) Len() int           { return len(s) }
-func (s bySize) Less(i, j int) bool { return len(s[i].vals) > len(s[j].vals) }
-func (s bySize) Swap(i, j int)      { s[i], s[j] = s[j], s[i] }